requirements.txt
torch==2.0.0
transformers==4.30.2
faiss-cpu==1.7.3
#faiss-gpu==1.7.3
annoy==1.17.1
nltk==3.8.1
numpy==1.24.3
pymupdf==1.24.4
